+++ java/tenmo-client/src/main/java/com/techelevator/tenmo/models/Transfers.java
@@ -1,5 +1,7 @@
 package com.techelevator.tenmo.models;
 
+import java.math.BigDecimal;
+
 public class Transfers {
 	
 		private int transfer_id;
@@ -7,11 +9,11 @@ public class Transfers {
 		private int transfer_status_id;
 		private int account_from;
 		private int account_to;
-		private double amount;
+		private BigDecimal amount;
 		
 		public Transfers() {}
 		
-		public Transfers(int transfer_id, int transfer_type_id, int transfer_status_id, int account_from, int account_to, double amount) {
+		public Transfers(int transfer_id, int transfer_type_id, int transfer_status_id, int account_from, int account_to, BigDecimal amount) {
 			this.transfer_id = transfer_id;
 			this.transfer_type_id = transfer_type_id;
 			this.transfer_status_id = transfer_status_id;
@@ -60,11 +62,11 @@ public class Transfers {
 			this.account_to = account_to;
 		}
 
-		public double getAmount() {
+		public BigDecimal getAmount() {
 			return amount;
 		}
 
-		public void setAmount(double amount) {
+		public void setAmount(BigDecimal amount) {
 			this.amount = amount;
 		}
